refactor: push hkdf to its own file
